+++ src/main/php/lang/ast/emit/PHP70.class.php
@@ -21,30 +21,19 @@ class PHP70 extends \lang\ast\Emitter {
   ];
 
   protected function emitCatch($catch) {
-    $last= array_pop($catch->types);
-    $label= sprintf('c%u', crc32($last));
-    foreach ($catch->types as $type) {
-      $this->out->write('catch('.$type.' $'.$catch->variable.') { goto '.$label.'; }');
+    if (empty($catch->types)) {
+      $this->out->write('catch(\\Throwable $'.$catch->variable.') {');
+    } else {
+      $last= array_pop($catch->types);
+      $label= sprintf('c%u', crc32($last));
+      foreach ($catch->types as $type) {
+        $this->out->write('catch('.$type.' $'.$catch->variable.') { goto '.$label.'; }');
+      }
+      $this->out->write('catch('.$last.' $'.$catch->variable.') { '.$label.':');
     }
 
-    $this->out->write('catch('.$last.' $'.$catch->variable.') { '.$label.':');
     $this->emit($catch->body);
     $this->out->write('}');
-  }
-
-  protected function emitAssignment($assignment) {
-    if ('array' === $assignment->variable->kind) {
-      $this->out->write('list(');
-      foreach ($assignment->variable->value as $pair) {
-        $this->emit($pair[1]);
-        $this->out->write(',');
-      }
-      $this->out->write(')');
-      $this->out->write($assignment->operator);
-      $this->emit($assignment->expression);
-    } else {
-      parent::emitAssignment($assignment);
-    }
   }
 
   protected function emitConst($const) {