+++ src/test/php/lang/ast/unittest/loader/Tests.php
@@ -1,5 +1,0 @@
-<?php namespace lang\ast\unittest\loader;
-
-class Tests {
-
-}